them giao dien menu

+++ src/main/java/geometry/Point.java
@@ -68,7 +68,7 @@ public class Point {
         return distance;
     }
 
-    public void setLocation(int x, int y) {
+    public void setLocation(double x, double y) {
         this.x = x;
         this.y = y;
     }
